+++ aim/engine/representation.py
@@ -1,7 +1,0 @@
-
-class ModelRepresentation:
-    def __init__(self):
-        pass
-
-    def get_ir(self):
-        pass